pyproject.toml
[tool.poetry]
name = "video-ghost"
version = "0.1.0"
description = ""
authors = ["Noir_Egoiste <dmitreykyrchin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
email-validator = "^2.1.1"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
